+++ packages/core-js/actual/promise/try.js
@@ -1,13 +1,14 @@
 'use strict';
 require('../../modules/es.promise');
 require('../../modules/esnext.promise.try');
-var call = require('../../internals/function-call');
+var apply = require('../../internals/function-apply');
 var isCallable = require('../../internals/is-callable');
 var path = require('../../internals/path');
 
 var Promise = path.Promise;
 var promiseTry = Promise['try'];
 
-module.exports = { 'try': function (callbackfn) {
-  return call(promiseTry, isCallable(this) ? this : Promise, callbackfn);
-} }['try'];
+// eslint-disable-next-line no-unused-vars -- required for arity
+module.exports = ({ 'try': function (callbackfn /* , ...args */) {
+  return apply(promiseTry, isCallable(this) ? this : Promise, arguments);
+} })['try'];